Refactor code, add possibility to send list of images for prediction
Refactored code to make it cleaner and more intuitive. Replaced model used by optimised version (gldv2_clean) together with config parameters.
Enabled user to send a list of images to exposed functions (local and global). Note: this is not batch processing yet. It will be done in the next step.